# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "An API for GreaseMeter, the cheesesteak rating app",
        "title": "GreaseMeter's Monolithic Rest API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/bookmarks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets bookmarks for user given a valid access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookmarks"
                ],
                "summary": "Gets user's bookmarks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Bookmark"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a users bookmark given a valid access token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookmarks"
                ],
                "summary": "Deletes a users bookmark",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/bookmarks/place/{id}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a valid user's bookmark for a place",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookmarks"
                ],
                "summary": "Creates a bookmark for a place",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Place ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/bookmarks/places/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets a user's bookmark status of a place",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookmarks"
                ],
                "summary": "Gets user's bookmark for place",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Place ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                }
            }
        },
        "/v1/places/list": {
            "get": {
                "description": "Gets list of places within a specified region",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "places"
                ],
                "summary": "Gets list of places",
                "parameters": [
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Latitude",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Longitude",
                        "name": "lng",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Latitude delta",
                        "name": "latDelta",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Longitude delta",
                        "name": "lngDelta",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page length",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.MetaPage"
                        }
                    }
                }
            }
        },
        "/v1/places/map": {
            "get": {
                "description": "Gets all map markers within a specified region",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "places"
                ],
                "summary": "Gets all map markers",
                "parameters": [
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Latitude",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Longitude",
                        "name": "lng",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Latitude delta",
                        "name": "latDelta",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Longitude delta",
                        "name": "lngDelta",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Marker"
                            }
                        }
                    }
                }
            }
        },
        "/v1/places/search": {
            "get": {
                "description": "Gets search results given a search term, and point",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "places"
                ],
                "summary": "Gets search results for places",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "term",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Latitude",
                        "name": "lat",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "format": "float64",
                        "description": "Longitude",
                        "name": "lng",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.SearchResult"
                            }
                        }
                    }
                }
            }
        },
        "/v1/places/{id}/images": {
            "get": {
                "description": "Gets images for a place given a valid place ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "places"
                ],
                "summary": "Gets images for a place",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Place ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/places/{id}/info": {
            "get": {
                "description": "Gets info for a place given a valid place ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "places"
                ],
                "summary": "Gets info for a place",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Place ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.PlaceInfo"
                        }
                    }
                }
            }
        },
        "/v1/places/{id}/meta": {
            "get": {
                "description": "Gets metadata for a place given a valid place ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "places"
                ],
                "summary": "Gets metadata for a place",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Place ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.PlaceMeta"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "types.Bookmark": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "placeId": {
                    "type": "integer"
                }
            }
        },
...
        "types.GeoJSONPoint": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "types.Marker": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "point": {
                    "$ref": "#/definitions/types.GeoJSONPoint"
                }
            }
        },
        "types.MetaPage": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.PlaceMeta"
                    }
                },
                "more": {
                    "type": "boolean"
                }
            }
        },
        "types.PlaceInfo": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rating": {
                    "type": "number"
                }
            }
        },
        "types.PlaceMeta": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                }
            }
        },
...
        "types.SearchResult": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}